requirements.txt
thefuzz
tqdm
jsonlines
pandas
python-dotenv
openai
numpy
orjson
numba
scipy
transformers==4.52.4